IOS-3699 Refactoring
